+++ chapter/51_腳手架自帶文件/public/manifest.json
