+++ chapter/react_staging/public/index.html
@@ -2,48 +2,15 @@
 <html lang="en">
 
 <head>
-  <meta charset="utf-8" />
-  <!-- %PUBLIC_URL% 代表 public 文件夾的路徑 -->
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>react腳手架</title>
   <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-  <!-- 開啟理想視口，用於做移動端網頁的適配 -->
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <!-- 用於配置瀏覽器頁籤 + 地址欄的顏色(僅支持安卓手機瀏覽器) -->
-  <meta name="theme-color" content="#000000" />
-  <meta name="description" content="Web site created using create-react-app" />
-  <!-- 用於指定網頁添加到手機主屏幕後的圖標 -->
-  <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-  <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-  <!-- 應用加殼時的配置文件 -->
-  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-  <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
-
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
-  <title>React App</title>
 </head>
 
 <body>
-  <!-- 若瀏覽器不支持 js 則展示標籤中的內容 -->
-  <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="root"></div>
-  <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
 </body>
 
 </html>
