+++ chapter/react_staging/src/Component/Hello/index.jsx
@@ -1,0 +1,8 @@
+import React, { Component } from "react";
+import hello from "./index.module.css"
+
+export default class Hello extends Component {
+    render() {
+        return <h2 className={hello.title}>Hello, React!</h2>
+    }
+}
