+++ chapter/react_staging/src/Component/Welcome/index.css
@@ -1,0 +1,3 @@
+.title {
+    background-color: skyblue;
+}
