+++ chapter/react_staging/src/Component/Welcome/index.jsx
@@ -1,0 +1,8 @@
+import React, { Component } from "react"
+import "./index.css"
+
+export default class Welcome extends Component {
+    render() {
+        return <h2 className="title">Welcome</h2>
+    }
+}
