+++ chapter/react_staging/src/index.js
@@ -1,17 +1,9 @@
+// 引入 react 核心庫
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+// 引入 ReactDOM 
+import ReactDOM from 'react-dom';
+// 引入 App 組件
+import App from "./App"
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+// 渲染 App 到頁面
+ReactDOM.render(<App />, document.getElementById("root"))
